+++ main.c
@@ -1114,27 +1114,23 @@ focus_direction(enum Direction direction) {
         switch (direction) {
         case DIRECTION_LEFT:
             dist = selected->x - client_aux->x - client_aux->w;
-            client_score = (int32)MIN(abs(dist),
-                                      abs(dist + selected->monitor->win_w));
+            client_score = MIN(abs(dist), abs(dist + selected->monitor->win_w));
             client_score += abs(selected->y - client_aux->y) - 1;
             break;
         case DIRECTION_RIGHT:
             dist = client_aux->x - selected->x - selected->w;
-            client_score = (int32)MIN(abs(dist),
-                                      abs(dist + selected->monitor->win_w));
+            client_score = MIN(abs(dist), abs(dist + selected->monitor->win_w));
             client_score += abs(client_aux->y - selected->y);
             break;
         case DIRECTION_UP:
             dist = selected->y - client_aux->y - client_aux->h;
-            client_score = (int32)MIN(abs(dist),
-                                      abs(dist + selected->monitor->win_h));
+            client_score = MIN(abs(dist), abs(dist + selected->monitor->win_h));
             client_score += abs(selected->x - client_aux->x) - 1;
             break;
         case DIRECTION_DOWN:
         default:
             dist = client_aux->y - selected->y - selected->h;
-            client_score = (int32)MIN(abs(dist),
-                                      abs(dist + selected->monitor->win_h));
+            client_score = MIN(abs(dist), abs(dist + selected->monitor->win_h));
             client_score += abs(client_aux->x - selected->x);
             break;
         }
